+++ PF_HashTable.h
@@ -10,12 +10,10 @@
 #include <list>
 
 struct PF_HashElement {
-//    PF_HashEntry *next;
-//    PF_HashEntry *prev;
-    int fd;
+    uintptr_t fd;
     PageNum pageNum;
     int slot;
-    PF_HashElement(int _fd, PageNum _pageNum, int _slot):fd(_fd), pageNum(_pageNum), slot(_slot) {}
+    PF_HashElement(uintptr_t _fd, PageNum _pageNum, int _slot):fd(_fd), pageNum(_pageNum), slot(_slot) {}
 };
 
 typedef std::list<PF_HashElement> PF_HashEntry;
@@ -24,15 +22,15 @@ class PF_HashTable {
 public:
     PF_HashTable(int numBuckets);
     ~PF_HashTable();
-    RC Find(int fd, PageNum pageNum, int &slot);
-    RC Insert(int fd, PageNum pageNum, int slot);
-    RC Delete(int fd, PageNum pageNum);
+    RC Find(uintptr_t fd, PageNum pageNum, int &slot);
+    RC Insert(uintptr_t fd, PageNum pageNum, int slot);
+    RC Delete(uintptr_t fd, PageNum pageNum);
 
 private:
     int numBuckets;
     std::vector<PF_HashEntry> hashTable;
-    int Hash(int fd, PageNum pageNum) const {
-        return (fd + pageNum) % numBuckets;
+    int Hash(uintptr_t fd, PageNum pageNum) const {
+        return (int)((fd + pageNum) % numBuckets);
     }
 };
